+++ example/src/main/java/ch/sbb/compose_mds/example/MainActivity.kt
@@ -37,7 +37,7 @@ import ch.sbb.compose_mds.example.pages.IconPage
 import ch.sbb.compose_mds.example.pages.LoadingIndicatorPage
 import ch.sbb.compose_mds.example.pages.MainPage
 import ch.sbb.compose_mds.example.pages.MessagePage
-import ch.sbb.compose_mds.example.pages.ModalViewPage
+import ch.sbb.compose_mds.example.pages.BottomSheetPage
 import ch.sbb.compose_mds.example.pages.NotificationBoxPage
 import ch.sbb.compose_mds.example.pages.RadioButtonPage
 import ch.sbb.compose_mds.example.pages.SegmentedButtonPage
@@ -88,7 +88,7 @@ class MainActivity : ComponentActivity() {
                             composable("switch") { SwitchPage() }
                             composable("header") { HeaderPage() }
                             composable("header-box") { HeaderBoxPage() }
-                            composable("modal-view") { ModalViewPage() }
+                            composable("bottom-sheet") { BottomSheetPage() }
                             composable("text-input") { TextInputPage() }
                             composable("textarea") { TextareaPage() }
                             composable("dropdown") { DropdownPage() }