+++ CHANGELOG.md
@@ -11,6 +11,8 @@ TBD
 * **SBBTextInput:** renamed from SBBTextField
 * **SBBTextInput:** refactored API and component
 * **SBBTextarea:** added component
+* **SBBModalView:** renamed to SBBBottomSheet
+* **SBBBottomSheet:** finished component, added isClosable and full-screen mode
 * **SBBDropdown:** added component
 * **SBBListItem:** added textStyle property and made trailingIcon optional
 * **SBBModalView:** use correct title font style